Update the IRepositorie and DeletedMovieHandler

+++ Galaxi.Movie.Domain/Handlers/DeletedMovieHandler.cs
@@ -5,7 +5,7 @@ using MediatR;
 namespace Galaxi.Movie.Domain.Handlers
 {
     public class DeletedMovieHandler
-        : IRequestHandler<DeleteMovieCommand, bool>
+        : IRequestHandler<DeleteMovieCommand, Unit>
     {
         private readonly IMovieRepository _repo;
 
@@ -13,11 +13,23 @@ namespace Galaxi.Movie.Domain.Handlers
         {
             _repo = repo;
         }
-        public async Task<bool> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteMovieCommand request, CancellationToken cancellationToken)
         {
             var function = await _repo.GetMovieById(request.FilmId);
+            if (function == null)
+            {
+                throw new DirectoryNotFoundException($"Movie with ID {request.FilmId} not found.");
+                
+            }
+
             _repo.Delete(function);
-            return await _repo.SaveAll();
+            var sucess = await _repo.SaveAll();
+            if (sucess)
+            {
+                throw new Exception("Failed to save changes.");
+            }
+
+            return Unit.Value;
         }
     }
 }
